+++ packages/theme/pigeonposse/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @dovenv/theme-pigeonposse
 
+## 1.5.29
+
+### Patch Changes
+
+- add feats to theme-pigeonposse, add dedent and other fuctions to utils
+
+- Updated dependencies []:
+  - @dovenv/stylelint-config@1.5.29
+  - @dovenv/eslint-config@1.5.29
+  - @dovenv/theme-banda@1.5.29
+  - @dovenv/core@1.5.29
+
 ## 1.5.28
 
 ### Patch Changes